# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Image MIME type, e.g. image/jpeg or image/png
 */
export type AnalyzeImageRequestMimeType =
  (typeof AnalyzeImageRequestMimeType)[keyof typeof AnalyzeImageRequestMimeType];

export const AnalyzeImageRequestMimeType = {
  "image/jpeg": "image/jpeg",
  "image/png": "image/png",
  "image/webp": "image/webp",
} as const;

/**
 * Optional body region the photo was taken of
 */
export type AnalyzeImageRequestBodyArea =
  (typeof AnalyzeImageRequestBodyArea)[keyof typeof AnalyzeImageRequestBodyArea];

export const AnalyzeImageRequestBodyArea = {
  general: "general",
  face: "face",
  scalp: "scalp",
  body: "body",
  hands: "hands",
  feet: "feet",
} as const;

export interface AnalyzeImageRequest {
  /**
   * Base64-encoded image data (without data URI prefix)
   * @minLength 32
   */
  imageBase64: string;
  /** Image MIME type, e.g. image/jpeg or image/png */
  mimeType: AnalyzeImageRequestMimeType;
  /** Optional body region the photo was taken of */
  bodyArea?: AnalyzeImageRequestBodyArea;
}

/**
 * Classification of the detected skin condition
 */
export type AnalysisResultCondition =
  (typeof AnalysisResultCondition)[keyof typeof AnalysisResultCondition];

export const AnalysisResultCondition = {
  none: "none",
  acne: "acne",
  benign_lesion: "benign_lesion",
  malignant_skin_cancer: "malignant_skin_cancer",
  eczema: "eczema",
  fungal_infection: "fungal_infection",
  other: "other",
} as const;

/**
 * Severity of the detected condition
 */
export type AnalysisResultSeverity =
  (typeof AnalysisResultSeverity)[keyof typeof AnalysisResultSeverity];

export const AnalysisResultSeverity = {
  none: "none",
  mild: "mild",
  moderate: "moderate",
  severe: "severe",
} as const;

/**
 * How urgently the user should seek medical care
 */
export type AnalysisResultUrgency =
  (typeof AnalysisResultUrgency)[keyof typeof AnalysisResultUrgency];

export const AnalysisResultUrgency = {
  routine: "routine",
  soon: "soon",
  urgent: "urgent",
} as const;

export type AnalysisResultDetailsItem = {
  label: string;
  value: string;
};

/**
 * Tailored health guidance for the detected condition
 */
export type AnalysisResultHealthGuide = {
  /** One-paragraph plain-language explanation of the condition */
  overview: string;
  /** Helpful self-care actions to take */
  dos: string[];
  /** Things to avoid */
  donts: string[];
  /** Warning signs that warrant a clinician visit */
  whenToSeeDoctor: string[];
};

export interface AnalysisResult {
  /** Short human-readable name for the detected condition (e.g. "Mild acne", "Healthy skin") */
  title: string;
  /** Classification of the detected skin condition */
  condition: AnalysisResultCondition;
  /** Whether an infection or skin condition is present in the image */
  infectionPresent: boolean;
  /** Severity of the detected condition */
  severity: AnalysisResultSeverity;
  /** How urgently the user should seek medical care */
  urgency: AnalysisResultUrgency;
  /**
   * Model confidence between 0 and 1
   * @minimum 0
   * @maximum 1
   */
  confidence: number;
  /** Plain-language summary of what was observed in the image */
  summary: string;
  /** Visible symptoms detected (redness, scaling, papules, etc.) */
  symptoms: string[];
  /** Notable observed attributes (location, color, texture, distribution) */
  details: AnalysisResultDetailsItem[];
  /** Tailored health guidance for the detected condition */
  healthGuide: AnalysisResultHealthGuide;
  /** Medical disclaimer reminding this is informational only */
  disclaimer: string;
}

export interface ErrorResponse {
  error: string;
}