# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultCondition } from "./analysisResultCondition";
import type { AnalysisResultDetailsItem } from "./analysisResultDetailsItem";
import type { AnalysisResultHealthGuide } from "./analysisResultHealthGuide";
import type { AnalysisResultSeverity } from "./analysisResultSeverity";
import type { AnalysisResultUrgency } from "./analysisResultUrgency";

export interface AnalysisResult {
  /** Short human-readable name for the detected condition (e.g. "Mild acne", "Healthy skin") */
  title: string;
  /** Classification of the detected skin condition */
  condition: AnalysisResultCondition;
  /** Whether an infection or skin condition is present in the image */
  infectionPresent: boolean;
  /** Severity of the detected condition */
  severity: AnalysisResultSeverity;
  /** How urgently the user should seek medical care */
  urgency: AnalysisResultUrgency;
  /**
   * Model confidence between 0 and 1
   * @minimum 0
   * @maximum 1
   */
  confidence: number;
  /** Plain-language summary of what was observed in the image */
  summary: string;
  /** Visible symptoms detected (redness, scaling, papules, etc.) */
  symptoms: string[];
  /** Notable observed attributes (location, color, texture, distribution) */
  details: AnalysisResultDetailsItem[];
  /** Tailored health guidance for the detected condition */
  healthGuide: AnalysisResultHealthGuide;
  /** Medical disclaimer reminding this is informational only */
  disclaimer: string;
}