# lib/api-zod/src/generated/types/analyzeImageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeImageRequestBodyArea } from "./analyzeImageRequestBodyArea";
import type { AnalyzeImageRequestMimeType } from "./analyzeImageRequestMimeType";

export interface AnalyzeImageRequest {
  /**
   * Base64-encoded image data (without data URI prefix)
   * @minLength 32
   */
  imageBase64: string;
  /** Image MIME type, e.g. image/jpeg or image/png */
  mimeType: AnalyzeImageRequestMimeType;
  /** Optional body region the photo was taken of */
  bodyArea?: AnalyzeImageRequestBodyArea;
}